Setting up Input Pt1
going to work with input handler,
want to really keep this organized this time, I added Sirenex's odin to unity and want to keep everything very customizeable and clean.

Project Goal: Create a custom sandbox 2d engine base that I create and manage ideas, game versions and states, and levels and puzzles with. This requires easy access and serialized component parts to quickly edit and play test iterations for a 2d platformer.