input overhaul and refactoring before adding more code
Now I am going to work in another scene and start on new scripts to refactor what I did in my current scripts. I am not taking advantage of delegates and the proper component framework to make this as modular as Id like. I really found that out when I added ps4 controller functionality. It could be the controller but seems to work fine on snes emulator.