holdableR and object 2D and Box Raycast System
fleshed out object 2d to be in space but it just hangs in air and isnt colliding with ground yet. It also isnt colliding with person, whcih is fine collisions can be done at holdable stage of abstraction.
]realized i can maybe hack together a solution using unity's gravity for everything but character, Im going to kinda explore that while fleshing out the non unity physics. It doesnt seem to optimal in anyway honestly and Im better off just keep working on 2D objects but I think It would be nice to have both.

I added a little dialogue when cahracter switches it explains which game object is switching. I eventually want it to objects that are created in the scene and know their name, but for now its premade.

also added boxcast to raycast system just in case i need it for holdable.
HoldableR is for rigidbody object but raycast player.

I cant wait to get out of this experimental code and into version 4. just want to get back to where i was in gameplay made in a version that couldnt grow. Also maybe ill add some little things along way